+++ database/migrations/2026_01_16_000002_create_simpatisan_drafts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('simpatisan_drafts', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('simpatisan_id')->nullable();
+            $table->unsignedInteger('simpatisan_id')->nullable();
             $table->enum('operation_type', ['create', 'update', 'delete']);
             $table->json('draft_data');
             $table->enum('status', ['draft', 'pending_review', 'approved', 'rejected', 'revision_required'])->default('draft');
